requirements/base.txt
Django==1.7
django-braces==1.4.0
django-model-utils==2.0.3
logutils==0.3.3
django-crispy-forms